fix(tier4_autoware_utils): fix `-Werror=float-conversion` (#5301)
fix(tier4_autoware_utils): fix

+++ common/tier4_autoware_utils/include/tier4_autoware_utils/geometry/geometry.hpp
@@ -245,7 +245,7 @@ inline void setOrientation(const geometry_msgs::msg::Quaternion & orientation, T
 }
 
 template <class T>
-void setLongitudinalVelocity([[maybe_unused]] const double velocity, [[maybe_unused]] T & p)
+void setLongitudinalVelocity([[maybe_unused]] const float velocity, [[maybe_unused]] T & p)
 {
   static_assert(sizeof(T) == 0, "Only specializations of getLongitudinalVelocity can be used.");
   throw std::logic_error("Only specializations of getLongitudinalVelocity can be used.");
@@ -253,14 +253,14 @@ void setLongitudinalVelocity([[maybe_unused]] const double velocity, [[maybe_unu
 
 template <>
 inline void setLongitudinalVelocity(
-  const double velocity, autoware_auto_planning_msgs::msg::TrajectoryPoint & p)
+  const float velocity, autoware_auto_planning_msgs::msg::TrajectoryPoint & p)
 {
   p.longitudinal_velocity_mps = velocity;
 }
 
 template <>
 inline void setLongitudinalVelocity(
-  const double velocity, autoware_auto_planning_msgs::msg::PathPoint & p)
+  const float velocity, autoware_auto_planning_msgs::msg::PathPoint & p)
 {
   p.longitudinal_velocity_mps = velocity;
 }
@@ -502,9 +502,9 @@ inline geometry_msgs::msg::Point32 transformPoint(
     geometry_msgs::build<geometry_msgs::msg::Point>().x(point32.x).y(point32.y).z(point32.z);
   const auto transformed_point = tier4_autoware_utils::transformPoint(point, pose);
   return geometry_msgs::build<geometry_msgs::msg::Point32>()
-    .x(transformed_point.x)
-    .y(transformed_point.y)
-    .z(transformed_point.z);
+    .x(static_cast<float>(transformed_point.x))
+    .y(static_cast<float>(transformed_point.y))
+    .z(static_cast<float>(transformed_point.z));
 }
 
 template <class T>
